% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicci贸n.R
\name{predicci贸n}
\alias{predicci贸n}
\title{Predicciones de regresion lineal multiple.}
\usage{
predicci贸n(Xn, ajuste)
}
\arguments{
\item{Xn}{(data.frame) de los datos de las variables explicativas.}

\item{ajuste}{es la salida de valores de la funcion Regresion_Lineal_M}
}
\value{
los valores de las predicciones de y
}
\description{
Realiza predicciones de una regresion linal multiple.
}
