% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regre_Lin_Multi.R
\name{Regresion_Lineal_M}
\alias{Regresion_Lineal_M}
\title{Regresion lineal multiple}
\usage{
Regresion_Lineal_M(X, Y)
}
\arguments{
\item{X}{(data.frame) de todos los datos.}

\item{Y}{(data.frame) de la variable respuesta.}
}
\value{
Una lista con las Betas, Minimo,Maximo, Error residual estandar, r^2 y r ajustada.
}
\description{
Realiza una regresion linal multiple.
}
